requirements.txt
django
djangorestframework
psycopg2
Pillow
django-extensions